+++ src/entities/post/post.ts
