+++ src/features/comment-management/ui/AddCommentDialog.tsx
@@ -1,6 +1,6 @@
-import { Dialog, DialogContent, DialogHeader, DialogTitle, Textarea, Button } from "../shared/ui"
+import { Dialog, DialogContent, DialogHeader, DialogTitle, Textarea, Button } from "../../../shared/ui"
 
-interface AddCommentDialogProps {
+interface AddCommentDialogProps { 
   isOpen: boolean
   onOpenChange: (open: boolean) => void
   newComment: {
@@ -29,7 +29,7 @@ const AddCommentDialog = ({
           <Textarea
             placeholder="댓글 내용"
             value={newComment.body}
-            onChange={(e) => onNewCommentChange("body", e.target.value)}
+            onChange={(e: React.ChangeEvent<HTMLTextAreaElement>) => onNewCommentChange("body", e.target.value)}
           />
           <Button onClick={onAddComment}>댓글 추가</Button>
         </div>
