+++ src/features/comment-management/ui/CommentList.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { Button } from "../shared/ui"
+import { Button } from "../../../shared/ui"
 import { Edit2, Plus, ThumbsUp, Trash2 } from "lucide-react"
 
 interface Comment {
