+++ src/features/comment-management/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as EditCommentDialog } from "./EditCommentDialog"
+export { default as AddCommentDialog } from "./AddCommentDialog"
+export { default as CommentList } from "./CommentList"
