+++ src/features/post-management/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as EditPostDialog } from "./EditPostDialog"
+export { default as AddPostDialog } from "./AddPostDialog"
+export { default as PostDetailDialog } from "./PostDetailDialog"
