+++ src/features/user-management/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as UserModal } from "./UserModal"
