+++ src/pages/PostsManagerPage.tsx
@@ -1,21 +1,17 @@
 import { useEffect, useState } from "react"
 import { Plus } from "lucide-react"
 import { useLocation, useNavigate } from "react-router-dom"
-import { Post, User } from "../types/post"
+import { Post, User } from "../entities/post"
 import { Button, Card, CardContent, CardHeader, CardTitle } from "../shared/ui"
-import SearchBar from "../components/SearchBar"
-import TagFilter from "../components/TagFilter"
-import SortFilter from "../components/SortFilter"
-import SortOrderFilter from "../components/SortOrderFilter"
-import PostTable from "../components/PostTable"
-import Pagination from "../components/Pagination"
-import AddPostDialog from "../components/AddPostDialog"
-import AddCommentDialog from "../components/AddCommentDialog"
-import EditCommentDialog from "../components/EditCommentDialog"
-import UserModal from "../components/UserModal"
-import PostDetailDialog from "../components/PostDetailDialog"
-import EditPostDialog from "../components/EditPostDialog"
-import CommentList from "../components/CommentList"
+import { SearchBar } from "../widgets/search-bar"
+import { TagFilter } from "../widgets/tag-filter"
+import { SortFilter } from "../widgets/sort-filter"
+import { SortOrderFilter } from "../widgets/sort-order-filter"
+import { PostTable } from "../widgets/post-table"
+import { Pagination } from "../widgets/pagination"
+import { AddPostDialog, EditPostDialog, PostDetailDialog } from "../features/post-management/ui"
+import { AddCommentDialog, EditCommentDialog, CommentList } from "../features/comment-management/ui"
+import { UserModal } from "../features/user-management/ui"
 
 interface PostData {
   posts: Post[]
