+++ src/widgets/post-table/PostTable.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction } from "react"
 import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
-import { Button, Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../shared/ui"
-import { Post, User } from "../types/post"
+import { Button, Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../../shared/ui"
+import { Post, User } from "../../entities/post"
 import React from "react"
 
 interface PostTableProps {
