+++ src/widgets/post-table/index.ts
@@ -1,0 +1,1 @@
+export { default as PostTable } from "./PostTable"
