+++ src/widgets/search-bar/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchBar } from "./SearchBar"
