+++ src/widgets/sort-filter/SortFilter.tsx
@@ -1,4 +1,4 @@
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../shared/ui"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../shared/ui"
 
 interface SortFilterProps {
   sortBy: string
