+++ src/widgets/sort-filter/index.ts
@@ -1,0 +1,1 @@
+export { default as SortFilter } from "./SortFilter"
