+++ src/widgets/sort-order-filter/SortOrderFilter.tsx
@@ -1,4 +1,4 @@
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../shared/ui"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../shared/ui"
 
 interface SortOrderFilterProps {
   sortOrder: string
