+++ src/widgets/sort-order-filter/index.ts
@@ -1,0 +1,1 @@
+export { default as SortOrderFilter } from "./SortOrderFilter"
