+++ src/widgets/tag-filter/index.ts
@@ -1,0 +1,1 @@
+export { default as TagFilter } from "./TagFilter"
